parse custom tags
