+++ packages/language-server/src/language-service/model/settings.ts
@@ -35,7 +35,6 @@ export interface LanguageSettings {
 	validate: boolean
 	hover: boolean
 	completion: boolean
-	customTags: string[]
 	validationProvider: ValidationProvider
 	cfnLint: CFNLintSettings
 	workspaceRoot: string
@@ -45,7 +44,6 @@ export const getDefaultLanguageSettings = (): LanguageSettings => ({
 	validate: true,
 	hover: true,
 	completion: true,
-	customTags: [],
 	validationProvider: ValidationProvider["cfn-lint"],
 	cfnLint: {
 		path: "cfn-lint",
