+++ packages/language-server/src/language-service/parser/__tests__/custom-tags.test.ts
@@ -1,0 +1,31 @@
+import { CUSTOM_TAGS, TagKind } from "./../../model/custom-tags"
+import { parse as parseYaml } from ".."
+
+const optionsMapping: { [key in TagKind]: string } = {
+	scalar: "logicVariable",
+	sequence: ["", "\t - logicalVariable1", "\t - logicalVariable2"].join("\n"),
+	mapping: "{ val1: logicalVariable1, val2: logicalVariable2 }"
+}
+
+describe("custom tags parse", () => {
+	describe("smoke tests", () => {
+		const generateNode = (text: string) => {
+			return parseYaml(text).documents[0].root
+		}
+
+		CUSTOM_TAGS.forEach(tag => {
+			describe(tag.type, () => {
+				test(`${tag.type}`, () => {
+					const text = [
+						`property: ${tag.tag} `,
+						optionsMapping[tag.kind]
+					].join("")
+					const node = generateNode(text)
+
+					// expect(node.customTag).toBeDefined()
+					expect(node).toMatchSnapshot()
+				})
+			})
+		})
+	})
+})
