+++ packages/language-server/src/language-service/parser/json/localize.ts
@@ -1,0 +1,3 @@
+import * as nls from "vscode-nls"
+
+export default nls.loadMessageBundle()
