+++ packages/language-server/src/language-service/parser/json/string-ast-node.ts
@@ -1,0 +1,81 @@
+import { ASTNode, ISchemaCollector } from "./ast-node"
+import * as Json from "jsonc-parser"
+import { JSONSchema } from "../../jsonSchema"
+import { CustomTag } from "../../model/custom-tags"
+import { ValidationResult, ProblemSeverity } from "./validation-result"
+import localize from "./localize"
+
+export class StringASTNode extends ASTNode {
+	isKey: boolean
+	value: string
+
+	constructor(
+		parent: ASTNode,
+		name: Json.Segment,
+		isKey: boolean,
+		start: number,
+		end?: number,
+		customTag?: CustomTag
+	) {
+		super(parent, "string", name, start, end, customTag)
+		this.isKey = isKey
+		this.value = ""
+	}
+
+	getValue(): string {
+		return this.value
+	}
+
+	validate(
+		schema: JSONSchema,
+		validationResult: ValidationResult,
+		matchingSchemas: ISchemaCollector
+	): void {
+		if (!matchingSchemas.include(this)) {
+			return
+		}
+		super.validate(schema, validationResult, matchingSchemas)
+
+		if (schema.minLength && this.value.length < schema.minLength) {
+			validationResult.problems.push({
+				location: { start: this.start, end: this.end },
+				severity: ProblemSeverity.Warning,
+				message: localize(
+					"minLengthWarning",
+					"String is shorter than the minimum length of {0}.",
+					schema.minLength
+				)
+			})
+		}
+
+		if (schema.maxLength && this.value.length > schema.maxLength) {
+			validationResult.problems.push({
+				location: { start: this.start, end: this.end },
+				severity: ProblemSeverity.Warning,
+				message: localize(
+					"maxLengthWarning",
+					"String is longer than the maximum length of {0}.",
+					schema.maxLength
+				)
+			})
+		}
+
+		if (schema.pattern) {
+			const regex = new RegExp(schema.pattern)
+			if (!regex.test(this.value)) {
+				validationResult.problems.push({
+					location: { start: this.start, end: this.end },
+					severity: ProblemSeverity.Warning,
+					message:
+						schema.patternErrorMessage ||
+						schema.errorMessage ||
+						localize(
+							"patternWarning",
+							'String does not match the pattern of "{0}".',
+							schema.pattern
+						)
+				})
+			}
+		}
+	}
+}
