+++ packages/language-server/src/language-service/services/documentSymbols/index.ts
@@ -1,6 +1,6 @@
 "use strict"
 
-import * as Parser from "../../parser/jsonParser"
+import * as Parser from "../../parser/json"
 import { YAMLDocument } from "./../../parser/index"
 
 import {
